requirements.txt
fastapi==0.85.0
Fiona==1.8.21
GDAL==3.4.3
geojson==2.5.0
Pillow==9.2.0
pyproj==3.2.1
rasterio==1.3.2
Shapely==1.8.4
uvicorn==0.15.0
